Fix flash of unstyled content on page load
When index.html and project.html were initially loaded in the browser
or refreshed, unstyled html was visible for a half second. This is due
to Webpack's bundling process, which loads CSS with JS scrips.
Webpack was re-configured with new plugins that extract the css and
create a stylesheet. This prevents FOUC from occuring. In addition,
an autoprefixer plugin and media query packer where added to the
project.