Add work-in-progress project section, fix buttons
Previously, many buttons and anchor tags had images inside them.
When a button with an image nested inside it was clicked, Chrome grabbed
the image's class, NOT the button's class (even though the button is the
parent element) To fix this, all the images that were inside anchor
tags or buttons are now displayed with css's backgroud-image property.