Add other projects and removed AOS styles
During the production bundling process the imported AOS style sheet
is not being minified even though the file aos/dist/aos.css is minified.
Additionally, only a few of the AOS styles are used in my project.
So, I removed all of the unused AOS styles and imported the style sheet
through sass so that it was minified.